Add keep alive test case rikeda

+++ srcs/config/TryFilesDirectiveHandler.cpp
@@ -14,7 +14,7 @@ void TryFilesDirectiveHandler::setConfig() {
   LocationConfig& locationConfig = getLocationConfig();
   for (size_t i = 1; i < this->tokens_.size() - 1; i++) {
     if (Utils::isStartsWith(tokens_[i], "=")) {
-      int status_code;
+      size_t status_code;
       std::string value = tokens_[i].substr(1);
       if (!Utils::parseValue(value, status_code)) {
         std::cout << "Failed to parse as int" << std::endl;
